+++ lib/aasm/persistence/base.rb
@@ -6,7 +6,7 @@ module AASM
         base.extend ClassMethods
       end
 
-      # Returns the value of the aasm_column - called from <tt>aasm.current_state</tt>
+      # Returns the value of the aasm.attribute_name - called from <tt>aasm.current_state</tt>
       #
       # If it's a new record, and the aasm state column is blank it returns the initial state
       # (example provided here for ActiveRecord, but it's true for Mongoid as well):
@@ -33,7 +33,7 @@ module AASM
       #
       # This allows for nil aasm states - be sure to add validation to your model
       def aasm_read_state
-        state = send(self.class.aasm_column)
+        state = send(self.class.aasm.attribute_name)
         if new_record?
           state.blank? ? aasm.determine_state_name(self.class.aasm.initial_state) : state.to_sym
         else
@@ -42,41 +42,9 @@ module AASM
       end
 
       module ClassMethods
-        # Maps to the aasm_column in the database.  Defaults to "aasm_state".  You can write
-        # (example provided here for ActiveRecord, but it's true for Mongoid as well):
-        #
-        #   create_table :foos do |t|
-        #     t.string :name
-        #     t.string :aasm_state
-        #   end
-        #
-        #   class Foo < ActiveRecord::Base
-        #     include AASM
-        #   end
-        #
-        # OR:
-        #
-        #   create_table :foos do |t|
-        #     t.string :name
-        #     t.string :status
-        #   end
-        #
-        #   class Foo < ActiveRecord::Base
-        #     include AASM
-        #     aasm_column :status
-        #   end
-        #
-        # This method is both a getter and a setter
-        def aasm_column(column_name=nil)
-          if column_name
-            AASM::StateMachine[self].config.column = column_name.to_sym
-            # @aasm_column = column_name.to_sym
-          else
-            AASM::StateMachine[self].config.column ||= :aasm_state
-            # @aasm_column ||= :aasm_state
-          end
-          # @aasm_column
-          AASM::StateMachine[self].config.column
+        def aasm_column(attribute_name=nil)
+          warn "[DEPRECATION] aasm_column is deprecated. Use aasm.attribute_name instead"
+          aasm.attribute_name(attribute_name)
         end
       end # ClassMethods
 
@@ -90,7 +58,7 @@ module AASM
       if AASM::StateMachine[@klass].config.create_scopes && !@klass.respond_to?(name)
         if @klass.ancestors.map {|klass| klass.to_s}.include?("ActiveRecord::Base")
 
-          conditions = {"#{@klass.table_name}.#{@klass.aasm_column}" => name.to_s}
+          conditions = {"#{@klass.table_name}.#{@klass.aasm.attribute_name}" => name.to_s}
           if ActiveRecord::VERSION::MAJOR >= 3
             @klass.class_eval do
               scope name, lambda { where(conditions) }
@@ -101,7 +69,7 @@ module AASM
             end
           end
         elsif @klass.ancestors.map {|klass| klass.to_s}.include?("Mongoid::Document")
-          scope_options = lambda { @klass.send(:where, {@klass.aasm_column.to_sym => name.to_s}) }
+          scope_options = lambda { @klass.send(:where, {@klass.aasm.attribute_name.to_sym => name.to_s}) }
           @klass.send(:scope, name, scope_options)
         end
       end